+++ Example/HGCircularSlider/ClockViewController.swift
@@ -15,11 +15,12 @@ extension Date {
 
 class ClockViewController: UIViewController {
     
-    
+
     @IBOutlet weak var durationLabel: UILabel!
     @IBOutlet weak var bedtimeLabel: UILabel!
     @IBOutlet weak var wakeLabel: UILabel!
     @IBOutlet weak var rangeCircularSlider: RangeCircularSlider!
+    @IBOutlet weak var clockFormatSegmentedControl: UISegmentedControl!
     
     lazy var dateFormatter: DateFormatter = {
         let dateFormatter = DateFormatter()
@@ -51,6 +52,11 @@ class ClockViewController: UIViewController {
     }
     
     @IBAction func updateTexts(_ sender: AnyObject) {
+        
+        adjustValue(value: &rangeCircularSlider.startPointValue)
+        adjustValue(value: &rangeCircularSlider.endPointValue)
+
+        
         let bedtime = TimeInterval(rangeCircularSlider.startPointValue)
         let bedtimeDate = Date(timeIntervalSinceReferenceDate: bedtime)
         bedtimeLabel.text = dateFormatter.string(from: bedtimeDate)
@@ -65,5 +71,12 @@ class ClockViewController: UIViewController {
         durationLabel.text = dateFormatter.string(from: durationDate)
         dateFormatter.dateFormat = "hh:mm a"
     }
+    
+    func adjustValue(value: inout CGFloat) {
+        let minutes = value / 60
+        let adjustedMinutes =  ceil(minutes / 5.0) * 5
+        value = adjustedMinutes * 60
+    }
+
 }