+++ HGCircularSlider/Classes/CircularSlider.swift
@@ -24,28 +24,28 @@ open class CircularSlider: UIControl {
      * The default value is a transparent color.
      */
     @IBInspectable
-    open var diskFillColor: UIColor = UIColor.clear
+    open var diskFillColor: UIColor = .clear
     
     /**
      * The color shown for the unselected portion of the slider disk. (outside start and end values)
      * The default value of this property is the black color with alpha = 0.3.
      */
     @IBInspectable
-    open var diskColor: UIColor = UIColor.gray
+    open var diskColor: UIColor = .gray
     
     /**
      * The color shown for the selected track portion. (between start and end values)
      * The default value of this property is the tint color.
      */
     @IBInspectable
-    open var trackFillColor: UIColor = UIColor.clear
+    open var trackFillColor: UIColor = .clear
     
     /**
      * The color shown for the unselected track portion. (outside start and end values)
      * The default value of this property is the white color.
      */
     @IBInspectable
-    open var trackColor: UIColor = UIColor.white
+    open var trackColor: UIColor = .white
     
     /**
      * The width of the circular line
@@ -54,7 +54,31 @@ open class CircularSlider: UIControl {
      */
     @IBInspectable
     open var lineWidth: CGFloat = 5.0
-    
+
+    /**
+     * The width of the unselected track portion of the slider
+     *
+     * The default value of this property is 5.0.
+     */
+    @IBInspectable
+    open var backtrackLineWidth: CGFloat = 5.0
+
+    /**
+     * The shadow offset of the slider
+     *
+     * The default value of this property is .zero.
+     */
+    @IBInspectable
+    open var trackShadowOffset: CGPoint = .zero
+
+    /**
+     * The color of the shadow offset of the slider
+     *
+     * The default value of this property is .gray.
+     */
+    @IBInspectable
+    open var trackShadowColor: UIColor = .gray
+
     /**
      * The width of the thumb stroke line
      *
@@ -78,14 +102,14 @@ open class CircularSlider: UIControl {
      * The default value of this property is the groupTableViewBackgroundColor.
      */
     @IBInspectable
-    open var endThumbTintColor: UIColor = UIColor.groupTableViewBackground
+    open var endThumbTintColor: UIColor = .groupTableViewBackground
     
     /**
      * The stroke highlighted color of the end thumb
      * The default value of this property is blue
      */
     @IBInspectable
-    open var endThumbStrokeHighlightedColor: UIColor = UIColor.blue
+    open var endThumbStrokeHighlightedColor: UIColor = .blue
     
     /**
      * The color used to tint the stroke of the end thumb
@@ -94,7 +118,7 @@ open class CircularSlider: UIControl {
      * The default value of this property is red.
      */
     @IBInspectable
-    open var endThumbStrokeColor: UIColor = UIColor.red
+    open var endThumbStrokeColor: UIColor = .red
     
     /**
      * The image of the end thumb
@@ -105,6 +129,18 @@ open class CircularSlider: UIControl {
     open var endThumbImage: UIImage?
     
     // MARK: Accessing the Slider’s Value Limits
+    
+    /**
+     * Fixed number of rounds - how many circles has user to do to reach max value (like apple bedtime clock - which have 2)
+     * the default value if this property is 1
+     */
+    @IBInspectable
+    open var numberOfRounds: Int = 1 {
+        didSet {
+            assert(numberOfRounds > 0, "Number of rounds has to be positive value!")
+            setNeedsDisplay()
+        }
+    }
     
     /**
      * The minimum value of the receiver.
@@ -145,14 +181,14 @@ open class CircularSlider: UIControl {
      * The default value of this property is 0.5
      */
     open var endPointValue: CGFloat = 0.5 {
-        didSet {            
+        didSet {
             if oldValue == endPointValue {
                 return
             }
             if endPointValue > maximumValue {
                 endPointValue = maximumValue
             }
-            
+
             setNeedsDisplay()
         }
     }
@@ -200,7 +236,8 @@ open class CircularSlider: UIControl {
     internal func setup() {
         trackFillColor = tintColor
     }
-    
+
+
     // MARK: Drawing methods
     
     /**
@@ -211,7 +248,7 @@ open class CircularSlider: UIControl {
         
         drawCircularSlider(inContext: context)
         
-        let valuesInterval = Interval(min: minimumValue, max: maximumValue)
+        let valuesInterval = Interval(min: minimumValue, max: maximumValue, rounds: numberOfRounds)
         // get end angle from end value
         let endAngle = CircularSliderHelper.scaleToAngle(value: endPointValue, inInterval: valuesInterval) + CircularSliderHelper.circleInitialAngle
         
@@ -243,17 +280,12 @@ open class CircularSlider: UIControl {
      See superclass documentation
      */
     override open func continueTracking(_ touch: UITouch, with event: UIEvent?) -> Bool {
-        
         // the position of the pan gesture
         let touchPosition = touch.location(in: self)
-        
         let startPoint = CGPoint(x: bounds.center.x, y: 0)
-        let angle = CircularSliderHelper.angle(betweenFirstPoint: startPoint, secondPoint: touchPosition, inCircleWithCenter: bounds.center)
+        let value = newValue(from: endPointValue, touch: touchPosition, start: startPoint)
         
-        let interval = Interval(min: minimumValue, max: maximumValue)
-        let newValue = CircularSliderHelper.value(inInterval: interval, fromAngle: angle)
-        
-        endPointValue = newValue
+        endPointValue = value
         sendActions(for: .valueChanged)
         
         return true
@@ -266,4 +298,24 @@ open class CircularSlider: UIControl {
         sendActions(for: .editingDidEnd)
     }
 
+    // MARK: Utilities methods
+    internal func newValue(from oldValue: CGFloat, touch touchPosition: CGPoint, start startPosition: CGPoint) -> CGFloat {
+        let angle = CircularSliderHelper.angle(betweenFirstPoint: startPosition, secondPoint: touchPosition, inCircleWithCenter: bounds.center)
+        let interval = Interval(min: minimumValue, max: maximumValue, rounds: numberOfRounds)
+        let deltaValue = CircularSliderHelper.delta(in: interval, for: angle, oldValue: oldValue)
+        
+        var newValue = oldValue + deltaValue
+        let range = maximumValue - minimumValue
+        
+        if newValue > maximumValue {
+            newValue -= range
+        }
+        else if newValue < minimumValue {
+            newValue += range
+        }
+        return newValue
+    }
+    
+    
+    
 }